Fixed the chunked tests
Divided the put tests into put and chunked. Added th chunked inti and
clean bash scripts.
Updated the configurations of webserv and nginx for them to accept PUT.
