+++ code/configurations/fakenginx.conf
@@ -71,5 +71,10 @@ server {
 		root /unitTest/put/;
 		request_method PUT GET;
 	}
+
+	location /chunked/ {
+		root /unitTest/chunked/;
+		request_method PUT GET;
+	}
 }
