+++ tests/dockerFiles/nginx/nginx.conf
@@ -96,6 +96,10 @@ server {
 		dav_methods PUT;
 	}
 
+	location /chunked {
+		dav_methods PUT;
+	}
+
 	location /put/notAllowed {
 		dav_methods off;
 	}
