+++ tests/scripts/requestsTests/cleanChunkedTest.sh
@@ -1,0 +1,1 @@
+rm -rf chunked/
