+++ tests/scripts/requestsTests/initChunkedTest.sh
@@ -1,0 +1,6 @@
+mkdir put
+chmod 777 put/
+
+##-----------------------------------------------------------------------------------------------------##
+	mkdir chunked/
+	chmod 777 chunked/
