+++ tests/scripts/requestsTests/utils.mjs
@@ -8,6 +8,28 @@ export const COLOR_BLUE = "\x1b[34m"
 export const COLOR_MAGENTA = "\x1b[35m"
 export const COLOR_CYAN = "\x1b[36m"
 
+
+const characters ='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+const charactersLength = characters.length;
+
+function	randomInt(min, max)
+{
+	return (Math.floor(Math.random() * (max - min) + min));
+}
+
+export function	generateString(min, max)
+{
+    let result = ' ';
+	const length = randomInt(min, max);
+
+    for (let i = 0; i < length; i++)
+	{
+		const randomIndex = randomInt(0, charactersLength);
+        result += characters[randomIndex];
+    }
+    return result;
+}
+
 export function	printHeader(header)
 {
 	console.log("");
